+++ src/components/calendar/MainCalendar.tsx
@@ -10,7 +10,7 @@ import { useUpadateEventMutate } from '@/hooks/mutations/useUpadateEventMutate';
 import { CustomDateHeader } from '@/components/calendar/CustomDateHeader';
 import { useGetHolidays } from '@/hooks/queries/useGetHolidays';
 import { holidayStyle } from '@/lib/utils/calendarStyle';
-import CalendarPopOver from '@/components/calendar/CalendarPopOver';
+import CalendarPopOver from '@/components/calendar/popOver/CalendarPopOver';
 
 // 드래그 이벤트 타입
 interface Props {
